Jsonifier Release v0.9.96
* Added support for directly prettifying the Json data as it's written.
* Fixing a string parsing issue with sizes.
* Added the ARM-NEON implementation for various string operations.

+++ Include/jsonifier/ISA/CompareValues.hpp
@@ -31,23 +31,23 @@ namespace simd_internal {
 
 #if JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_NEON)
 
-	template<simd_int_128_type simd_int_t01> JSONIFIER_INLINE static uint32_t toBitMask(simd_int_t01&& value) {
+	template<simd_int_128_type simd_int_t01> JSONIFIER_INLINE static uint16_t toBitMask(simd_int_t01&& value) {
 		static constexpr uint8x16_t bit_mask{ 0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80, 0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80 };
 		auto minput	   = value & bit_mask;
 		uint8x16_t tmp = vpaddq_u8(minput, minput);
 		tmp			   = vpaddq_u8(tmp, tmp);
 		tmp			   = vpaddq_u8(tmp, tmp);
-		return vgetq_lane_u16(vreinterpretq_u16_u8(tmp), 0);
+		return static_cast<uint16_t>(vgetq_lane_u16(vreinterpretq_u16_u8(tmp), 0));
 	}
 
-	template<simd_int_128_type simd_int_t01, simd_int_128_type simd_int_t02> JSONIFIER_INLINE static uint32_t opCmpEq(simd_int_t01&& value, simd_int_t02&& other) {
+	template<simd_int_128_type simd_int_t01, simd_int_128_type simd_int_t02> JSONIFIER_INLINE static uint16_t opCmpEq(simd_int_t01&& value, simd_int_t02&& other) {
 		return toBitMask(vceqq_u8(value, other));
 	}
 
 #elif JSONIFIER_CHECK_FOR_AVX(JSONIFIER_AVX)
 
 	template<simd_int_128_type simd_int_t01, simd_int_128_type simd_int_t02> JSONIFIER_INLINE static auto opCmpEq(simd_int_t01&& value, simd_int_t02&& other) {
-		return static_cast<uint32_t>(_mm_movemask_epi8(_mm_cmpeq_epi8(std::forward<simd_int_t01>(value), std::forward<simd_int_t02>(other))));
+		return static_cast<uint16_t>(_mm_movemask_epi8(_mm_cmpeq_epi8(std::forward<simd_int_t01>(value), std::forward<simd_int_t02>(other))));
 	}
 
 	#if JSONIFIER_CHECK_FOR_AVX(JSONIFIER_AVX2)
@@ -68,8 +68,8 @@ namespace simd_internal {
 
 #else
 
-	template<simd_int_128_type simd_int_t01, simd_int_128_type simd_int_t02> JSONIFIER_INLINE static uint32_t opCmpEq(simd_int_t01&& value, simd_int_t02&& other) {
-		return static_cast<uint32_t>(_mm128_movemask_epi8(_mm128_cmpeq_epi8(std::forward<simd_int_t01>(value), std::forward<simd_int_t02>(other), std::make_index_sequence<16>{}),
+	template<simd_int_128_type simd_int_t01, simd_int_128_type simd_int_t02> JSONIFIER_INLINE static uint16_t opCmpEq(simd_int_t01&& value, simd_int_t02&& other) {
+		return static_cast<uint16_t>(_mm128_movemask_epi8(_mm128_cmpeq_epi8(std::forward<simd_int_t01>(value), std::forward<simd_int_t02>(other), std::make_index_sequence<16>{}),
 			std::make_index_sequence<16>{}));
 	}
